backend/dd_ml_segmentation_benchmark/requirements.txt
image-classifiers
keras
numpy==1.17.1
opencv_python==3.4.3.18
sklearn
tensorflow-gpu
torch
typing==3.6.6
wandb
webcolors
matplotlib
sklearn
scikit-image